pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-os"
version = "0.1.0"
description = "The open-source operations layer for AI agents."
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [
    { name = "Corvyd", email = "hello@corvyd.ai" },
]
keywords = ["ai", "agents", "operations", "multi-agent", "autonomous"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "claude-agent-sdk>=0.1.36",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
    "pre-commit>=3.6",
]

[project.urls]
Homepage = "https://github.com/corvyd-ai/agent-os"
Documentation = "https://github.com/corvyd-ai/agent-os/tree/main/docs"
Repository = "https://github.com/corvyd-ai/agent-os"
Issues = "https://github.com/corvyd-ai/agent-os/issues"

[project.scripts]
agent-os = "agent_os.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_os"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# --- Linting & formatting (ruff) ---

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade (modern Python syntax)
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",    # line length handled by formatter
    "SIM108",  # ternary operator — readability over brevity
    "B010",    # setattr with constant — used in test fixtures
    "E402",    # imports after try/except SDK fallbacks in runner.py
    "N806",    # fallback class assignments (CLIConnectionError = None)
    "N818",    # private sentinel _Unavailable is not a real exception
    "RUF012",  # mutable class attrs used as constants in ErrorClassifier
    "TC001",   # type-checking block — imports are used at runtime
    "TC003",   # type-checking block — stdlib imports used at runtime
]

[tool.ruff.lint.isort]
known-first-party = ["agent_os"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
